Kernel: Fix linker error
There was a weird error that wasn't allowing me to link, which was
super annoying. Anyways, it turned out to be that I was trying to
compile the kernel with PIE enabled, but the bootstrap wasn't, so
it wasn't able to link the two. This is, in retrospect, super obvious
and dumb, but it's a journey. Maybe I'll get better eventually.

Yet again, GCC finds even more ways to show just how stupid I really am.

P.S.
I don't have QEMU on this system yet, so I can't even test if the kernel
can even boot. Have fun with that, future me. :)