Kernel: Correct boot.S and add abstraction to write to VGA text buffer
boot.S:
I mistyped and accidentally put both pdts in low, physical memory and
not one low and one high, like I was supposed to. That's why when we
jumped into C++ it was trowing a memory access violation. The memory was
never marked as present.

Graphics/VGATextModeDevice:
This is an abstraction that allows us to easily write to the VGA text
buffer.
In the future, we should just write a different abstraction the draws
text to the framebuffer, rather than the VGA text buffer. However, we're
just going to keep using this for the time being until the kernel is in
a halfway-decent place. I'm just more internested in working on the
kernel than anything.