Toolchain: Further develop the toolchain.
I've been developing the toolchain so i could compile on my mac, but
that's quickly becoming more effort than it's worth, so I'll just
develop in a linux VM or on my desktop until i feel that I want to
finish it. It's just so boring and it's burning me out.