Axiom: Finish bootloader transition
The switch to limine is now complete, so we can now boot into long mode
easily since the bootloader does it for us. In addition to 64-bit long
mode support, limine also has UEFI capabilities, so it can boot in
either legacy BIOS mode or the more modern UEFI mode. Considering how
little passion I have for setting up these things on my own, I'm more
than happy to allow the bootloader to take care of it for me.

Next up on the docket is setting up a memory manager. Well... I suppose
we should set up an IDT first and do all the other setup stuff for the
CPU that needs to happen. Idk. It'll all depend on how I'm feeling when
I sit down to work on this again.

We're one step closer to the finished product :)