Axiom: Add a ton of new stuff
[Kernel]
We now have a way to print debug info to the host terminal, which can be
extremely helpful. This is accomplished by broadcasting the individual
bytes over the port 0xE9. and as long as you have the qemu option
-debugcon stdio set, then it will output all that data to your host
terminal.

[Toolchain]
We now have a funtional cross-compiler. Initially I was trying to an
OS-specific toolchain, but seeing as how I don't have literally
anything. It's probably not neccessary. So we just have a generic
x86_64-pc-elf target.