[finished] Kernel/Memory: Implement k_malloc
We now have a very basic implementation of k_malloc and k_free. It's
just a basic first-fit allocator that certainly isn't the best, but
it'll work for now and help us keep forward momentum and work on cool
new features.

The next goal will be to implement a memory manager. After that, we need
to get SMP support going and start on the scheduler so that we can start
jumping into the wonderful land of userspace.