AXUtil: Add `List` template
This commit adds a simple list template.

The inner workings are quite simple. It's a basic heap-allocated list
that, when you try to add beyond the list's current capacity, it just
doubles in size. It might not be ideal to simply double it's size, but I
want to prioritize less allocations, since my allocator probably isn't
very fast.