`AXUtil::Bitmap`/k_debug: Fix allocation bug and impove logging
`Bitmap`: Fix allocation bug
There was a major bug here where, when we were marking a range in the
bitmap, we would stop after the first chunk bit. This means that if we
ever allocated anything bigger than the chunk size, we would have
overlapping allocations. This was producing a huge headache in
`MemoryManager` because there were some data field in a list that were
getting overwritten by some of the elements of another list that came
just after it in memory. It turned out to be that I had just broke from
the loop that was setting all the bits of the bitmap right after setting
the first bit. I removed those breaks, and it now functions as-intended.

k_debug: Improve logging
we now have an improved logging system that allows different log levels
(info, warning, and error) as well as denoting the general task in which
they are occuring. I plan on modifying this general location feature to
be the actual name of the funtion in which the log function is called.
But that's not very important right now. So, we'll cross that bridge
when we come to it. :)