kernel/memory: Seperate the memory into it's individual components
This patch seperates out the memory manager into it's two components.
That of course being the physical manager, and the virtual manager.

This makes the whole system more manageable and readable.

Additional changes:
The `debug_text_console` has been renamed to the `early_console` since
that's what we'll boot into initially before we start the GUI. This will
make the initial testing stages easier so we know when things are
happening without having to look at the terminal output.